Extra package with 2 tests created - Test Base class contains beginning and the end of the test
